requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Deep learning (for Phase 2)
torch>=2.0.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Optional: Experiment tracking
# wandb>=0.15.0

# Optional: Advanced config management
# hydra-core>=1.3.0

# Optional: Visualization
# matplotlib>=3.7.0
# seaborn>=0.12.0
